refactor(sidebar): H2.2 — extract useSidebarNewReleasesUnread hook
